FixCSVOp-setCapMax-UserPassDB
Users and passwords data base is working correctly by interacting with the config.ini file. Reading and Writing to de .csv file are already fixed, now performing both actions when asked. Also the setCapacidadMaxima function was created and it's working correctly, changing the max capacity value when asked and replacing it, overriding the past value set on the config.ini file